moves reducer outside of store directory
