+++ src/App/Project/CreateNew/schema.js
@@ -1,7 +1,7 @@
 export const schema = [
   {
     fieldType: "input",
-    name: "projectName",
+    name: "name",
     ref: ref => ref({ required: true }),
     placeholder: "Project Name",
     error: "Project name is required."
