+++ src/index.js
@@ -5,7 +5,7 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 import { createStore, applyMiddleware } from "redux";
-import rootReducer from "./shared/store/reducers/rootReducer";
+import rootReducer from "./shared/reducers/rootReducer";
 import { Provider } from "react-redux";
 import thunk from "redux-thunk";
 
@@ -16,9 +16,9 @@ import thunk from "redux-thunk";
 const store = createStore(rootReducer, applyMiddleware(thunk));
 
 ReactDOM.render(
-    <Provider store={store}>
-        <App />
-    </Provider>,
-    document.getElementById("root")
+  <Provider store={store}>
+    <App />
+  </Provider>,
+  document.getElementById("root")
 );
 serviceWorker.register();
