+++ src/shared/reducers/rootReducer.js
@@ -1,10 +1,12 @@
 import authReducer from "./authReducer";
 import appReducer from "./appReducer";
+import dashboardReducer from "./dashboardReducer";
 import { combineReducers } from "redux";
 
 const rootReducer = combineReducers({
   app: appReducer,
-  auth: authReducer
+  auth: authReducer,
+  dashboard: dashboardReducer
 });
 
 export default rootReducer;
